Update _config.yml
Experimenting the configuration of my website: Github uses the 'index.html' file that is located in the master branch as content of my web page. I want to apply the Jekyll theme: slate to this web page. I want to see first if, by excluding 'index.html' from the 'config.yml' file that selects the theme, I get an empty web site with the slate theme.